fix(wallet): address CodeRabbit feedback on the lifecycle overlay
- managerForStoredWalletOperation shuts the detached manager down before
  rethrowing a loadFromPersistor failure (async ripple to managerForWipe
  callers) instead of leaking its native teardown to the deinit fallback
- the wipe's runtime-teardown wait is bounded (180 s): on expiry the wipe
  reports failure - never success while fullReset is unfinished - and the
  queued teardown keeps running in the background
- WalletLifecycleTransitionState and the overlay quartet (presenter,
  Obj-C bridge, view model, card view) move to their own files per the
  one-file-one-responsibility guideline; no behavior change
